# swagger.yaml
openapi: 3.0.3
info:
  title: Product Service API
  description: REST API for managing products (CRUD) using Node.js, Express, TypeORM, Redis, and GraphQL.
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
    description: Local development server

paths:
  /products:
    get:
      summary: Get all products
      description: Retrieve a list of all products.
      responses:
        "200":
          description: A list of products.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      summary: Create a new product
      description: Add a new product to the database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      summary: Get product by ID
      description: Retrieve a product by its unique ID.
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found

    put:
      summary: Update product by ID
      description: Modify an existing product’s details.
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      summary: Delete product by ID
      description: Remove a product from the database.
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
      responses:
        "204":
          description: Product deleted successfully
        "404":
          description: Product not found

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "iPhone 16"
        description:
          type: string
          example: "Apple iPhone 16 with A19 chip"
        price:
          type: number
          format: float
          example: 1299.99
        createdAt:
          type: string
          format: date-time
          example: "2025-10-21T14:48:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-10-21T15:00:00.000Z"

    ProductInput:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
          example: "iPhone 16"
        description:
          type: string
          example: "Latest Apple smartphone"
        price:
          type: number
          format: float
          example: 1299.99